src/requirements.txt
strands-agents==1.* 
strands-agents-tools>=0.2.7 

mcp==1.*

boto3==1.*
